Added handling for null platform values
Whenever HonkBot is unable to get a platform specific share url for a
song on YouTube, Spotify, or Apple Music, it will now disable the button
for that platform. This should fix the issue with HonkBot getting stuck
on "HonkBot is thinking..." when it can't find a share url for that
platform.